Split up tests into a file per test topic
